Remove Symfony UX Turbo and refactoring
Symfony UX Turbo dependency was removed from the project. Refactoring was performed on various files - AppFixtures was renamed to ActorFixtures, MoviePictures to MovieFixtures, and their corresponding data was updated. In MoviesController, entity manager was added to work with Movie entity and fetch all the movies. The changes made bring more clarity and better organization to the code.